removes a useless float/int conversion.

+++ polyora/keypoint.h
@@ -36,7 +36,7 @@ class keypoint
 {
  public:
    keypoint(void) { u=-1; v=-1; scale=0; id=-1; }
-   keypoint(float u, float v, float scale) { this->u = u; this->v = v; this->scale = scale; id=-1; }
+   keypoint(float u, float v, int scale) { this->u = u; this->v = v; this->scale = scale; id=-1; }
    keypoint(const keypoint & p) { u = p.u; v = p.v; scale = p.scale; score = p.score; id=p.id; }
    virtual ~keypoint() {}
